Fix flaky producer/consumer (#10)
Now both consumer and producers recovers from failures in their internals process preventing exceptions of being thrown by the flow it self and stoping it's execution. All exception are thrown by the loggers to ease debugging.

Producers use a custom flow that allow the reuse of materialize flows and implement backpressure on the generation of the message, allowing a finer grained control of the sent message.